+++ src/hooks/index.ts
@@ -3,5 +3,6 @@ import useHeader from "./useHeader.ts";
 import useLogin from "./useLogin.ts";
 import useModal from "./useModal.ts";
 import useToast from "./useToast.ts";
+import useToastHandler from "./useToastHandler.ts";
 
-export { useHamburger, useHeader, useLogin, useModal, useToast };
+export { useHamburger, useHeader, useLogin, useModal, useToast, useToastHandler };
